+++ config/configauth/configauth.go
@@ -28,20 +28,45 @@ var (
 
 // Authentication defines the auth settings for the receiver.
 type Authentication struct {
-	// AuthenticatorName specifies the name of the extension to use in order to authenticate the incoming data point.
-	AuthenticatorName string `mapstructure:"authenticator"`
+	// AuthenticatorID specifies the name of the extension to use in order to authenticate the incoming data point.
+	AuthenticatorID config.ComponentID `mapstructure:"authenticator"`
 }
 
-// GetServerAuthenticator attempts to select the appropriate from the list of extensions, based on the requested extension name.
-// If an authenticator is not found, an error is returned.
-func GetServerAuthenticator(extensions map[config.ComponentID]component.Extension, componentID config.ComponentID) (ServerAuthenticator, error) {
-	for id, ext := range extensions {
+// GetServerAuthenticator attempts to select the appropriate ServerAuthenticator from the list of extensions,
+// based on the requested extension name. If an authenticator is not found, an error is returned.
+func (a Authentication) GetServerAuthenticator(extensions map[config.ComponentID]component.Extension) (ServerAuthenticator, error) {
+	if ext, found := extensions[a.AuthenticatorID]; found {
 		if auth, ok := ext.(ServerAuthenticator); ok {
-			if id == componentID {
-				return auth, nil
-			}
+			return auth, nil
 		}
+		return nil, fmt.Errorf("requested authenticator is not a server authenticator")
 	}
 
-	return nil, fmt.Errorf("failed to resolve authenticator %q: %w", componentID.String(), errAuthenticatorNotFound)
+	return nil, fmt.Errorf("failed to resolve authenticator %q: %w", a.AuthenticatorID, errAuthenticatorNotFound)
+}
+
+// GetHTTPClientAuthenticator attempts to select the appropriate HTTPClientAuthenticator from the list of extensions,
+// based on the component id of the extension. If an authenticator is not found, an error is returned.
+// This should be only used by HTTP clients.
+func (a Authentication) GetHTTPClientAuthenticator(extensions map[config.ComponentID]component.Extension) (HTTPClientAuthenticator, error) {
+	if ext, found := extensions[a.AuthenticatorID]; found {
+		if auth, ok := ext.(HTTPClientAuthenticator); ok {
+			return auth, nil
+		}
+		return nil, fmt.Errorf("requested authenticator is not a HTTPClientAuthenticator")
+	}
+	return nil, fmt.Errorf("failed to resolve authenticator %q: %w", a.AuthenticatorID, errAuthenticatorNotFound)
+}
+
+// GetGRPCClientAuthenticator attempts to select the appropriate GRPCClientAuthenticator from the list of extensions,
+// based on the component id of the extension. If an authenticator is not found, an error is returned.
+// This should only be used by gRPC clients.
+func (a Authentication) GetGRPCClientAuthenticator(extensions map[config.ComponentID]component.Extension) (GRPCClientAuthenticator, error) {
+	if ext, found := extensions[a.AuthenticatorID]; found {
+		if auth, ok := ext.(GRPCClientAuthenticator); ok {
+			return auth, nil
+		}
+		return nil, fmt.Errorf("requested authenticator is not a GRPCClientAuthenticator")
+	}
+	return nil, fmt.Errorf("failed to resolve authenticator %q: %w", a.AuthenticatorID, errAuthenticatorNotFound)
 }