+++ service/collector_test.go
@@ -282,7 +282,7 @@ func TestCollector_reloadService(t *testing.T) {
 
 	tests := []struct {
 		name           string
-		parserProvider parserprovider.MapProvider
+		parserProvider config.MapProvider
 		service        *service
 	}{
 		{