+++ service/settings.go
@@ -21,7 +21,6 @@ import (
 	"go.opentelemetry.io/collector/component"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configunmarshaler"
-	"go.opentelemetry.io/collector/service/parserprovider"
 )
 
 // svcSettings holds configuration for building a new service.
@@ -63,8 +62,8 @@ type CollectorSettings struct {
 	// If it is not provided a default provider is used. The default provider loads the configuration
 	// from a config file define by the --config command line flag and overrides component's configuration
 	// properties supplied via --set command line flag.
-	// If the provider is parserprovider.Watchable, collector may reload the configuration upon error.
-	ConfigMapProvider parserprovider.MapProvider
+	// If the provider is config.WatchableMapProvider, collector may reload the configuration upon error.
+	ConfigMapProvider config.MapProvider
 
 	// ConfigUnmarshaler unmarshalls the configuration's Parser into the service configuration.
 	// If it is not provided a default unmarshaler is used.